Outsorced the function for the api
